zigbee: changed API to fix an inconsistence of extendedPanId java type between ZigBeeHost.setExtendedPanId and ZigBeeNode.getExtendedPanId.  Using always BigInteger.

+++ org.osgi.impl.service.zigbee/src/org/osgi/impl/service/zigbee/basedriver/ZigBeeHostImpl.java
@@ -75,7 +75,7 @@ public class ZigBeeHostImpl extends ZigBeeNodeImpl implements ZigBeeHost {
 
 	}
 
-	public void setExtendedPanId(long extendedPanId) {
+	public void setExtendedPanId(BigInteger extendedPanId) {
 
 	}
