+++ org.osgi.test.cases.zigbee/src/org/osgi/test/cases/zigbee/mock/ZigBeeHostImpl.java
@@ -72,7 +72,7 @@ public class ZigBeeHostImpl extends ZigBeeNodeImpl implements ZigBeeHost {
 
 	}
 
-	public void setExtendedPanId(long extendedPanId) {
+	public void setExtendedPanId(BigInteger extendedPanId) {
 
 	}
